refactor: split node proxy routes
