+++ server/node/server.cjs
@@ -7,12 +7,13 @@ const http = require('http');
 const path = require('path');
 const net = require('net');
 const htmlparser = require('node-html-parser');
-const { existsSync, readFileSync, writeFileSync } = require('fs');
+const { writeFileSync } = require('fs');
 const fs = require('fs/promises')
 const crypto = require('crypto')
 const { WebSocketServer } = require('ws');
 const { createAuthHelpers, createLimiters } = require('./auth.cjs');
-const { savePath, passwordPath, knownPublicKeysPath, authCodePath, sslPath, hubURL, proxyStreamConfig } = require('./config.cjs');
+const { authCodePath, sslPath, proxyStreamConfig } = require('./config.cjs');
+const { registerProxyRoutes } = require('./routes-proxy.cjs');
 const { registerStorageRoutes } = require('./routes-storage.cjs');
 const { createServerState } = require('./state.cjs');
 app.use(express.static(path.join(process.cwd(), 'dist'), {index: false}));
@@ -44,35 +45,6 @@ const authHelpers = {
     checkAuth,
 };
 const proxyStreamJobs = state.proxyStreamJobs;
-
-function getRequestTimeoutMs(timeoutHeader) {
-    const raw = Array.isArray(timeoutHeader) ? timeoutHeader[0] : timeoutHeader;
-    if (!raw) {
-        return null;
-    }
-    const timeoutMs = Number.parseInt(raw, 10);
-    if (!Number.isFinite(timeoutMs) || timeoutMs <= 0) {
-        return null;
-    }
-    return timeoutMs;
-}
-
-function createTimeoutController(timeoutMs) {
-    if (!timeoutMs) {
-        return {
-            signal: undefined,
-            cleanup: () => {}
-        };
-    }
-
-    const controller = new AbortController();
-    const timer = setTimeout(() => controller.abort(), timeoutMs);
-
-    return {
-        signal: controller.signal,
-        cleanup: () => clearTimeout(timer)
-    };
-}
 
 function normalizeProxyStreamTimeoutMs(timeoutMs) {
     if (!Number.isFinite(timeoutMs) || timeoutMs <= 0) {
@@ -499,317 +471,11 @@ app.get('/', async (req, res, next) => {
     }
 })
 
-const reverseProxyFunc = async (req, res, next) => {
-    if(!await checkProxyAuth(req, res)){
-        return;
-    }
-    
-    const urlParam = req.headers['risu-url'] ? decodeURIComponent(req.headers['risu-url']) : req.query.url;
-
-    if (!urlParam) {
-        res.status(400).send({
-            error:'URL has no param'
-        });
-        return;
-    }
-    const header = req.headers['risu-header'] ? JSON.parse(decodeURIComponent(req.headers['risu-header'])) : req.headers;
-    if(!header['x-forwarded-for']){
-        header['x-forwarded-for'] = req.ip
-    }
-
-    if(req.headers['authorization']?.startsWith('X-SERVER-REGISTER')){
-        if(!existsSync(authCodePath)){
-            delete header['authorization']
-        }
-        else{
-            const authCode = await fs.readFile(authCodePath, {
-                encoding: 'utf-8'
-            })
-            header['authorization'] = `Bearer ${authCode}`
-        }
-    }
-    const timeoutMs = getRequestTimeoutMs(req.headers['risu-timeout-ms']);
-    const timeout = createTimeoutController(timeoutMs);
-    let originalResponse;
-    try {
-        // make request to original server
-        originalResponse = await fetch(urlParam, {
-            method: req.method,
-            headers: header,
-            body: JSON.stringify(req.body),
-            signal: timeout.signal
-        });
-        // get response body as stream
-        const originalBody = originalResponse.body;
-        // get response headers
-        const head = new Headers(originalResponse.headers);
-        head.delete('content-security-policy');
-        head.delete('content-security-policy-report-only');
-        head.delete('clear-site-data');
-        head.delete('Cache-Control');
-        head.delete('Content-Encoding');
-        const headObj = {};
-        for (let [k, v] of head) {
-            headObj[k] = v;
-        }
-        // send response headers to client
-        res.header(headObj);
-        // send response status to client
-        res.status(originalResponse.status);
-        // send response body to client
-        await pipeline(originalResponse.body, res);
-
-    }
-    catch (err) {
-        if (err?.name === 'AbortError') {
-            if (!res.headersSent) {
-                res.status(504).send({
-                    error: timeoutMs
-                        ? `Proxy request timed out after ${timeoutMs}ms`
-                        : 'Proxy request aborted'
-                });
-            } else {
-                res.end();
-            }
-            return;
-        }
-        next(err);
-        return;
-    } finally {
-        timeout.cleanup();
-    }
-}
-
-const reverseProxyFunc_get = async (req, res, next) => {
-    if(!await checkProxyAuth(req, res)){
-        return;
-    }
-    
-    const urlParam = req.headers['risu-url'] ? decodeURIComponent(req.headers['risu-url']) : req.query.url;
-
-    if (!urlParam) {
-        res.status(400).send({
-            error:'URL has no param'
-        });
-        return;
-    }
-    const header = req.headers['risu-header'] ? JSON.parse(decodeURIComponent(req.headers['risu-header'])) : req.headers;
-    if(!header['x-forwarded-for']){
-        header['x-forwarded-for'] = req.ip
-    }
-    const timeoutMs = getRequestTimeoutMs(req.headers['risu-timeout-ms']);
-    const timeout = createTimeoutController(timeoutMs);
-    let originalResponse;
-    try {
-        // make request to original server
-        originalResponse = await fetch(urlParam, {
-            method: 'GET',
-            headers: header,
-            signal: timeout.signal
-        });
-        // get response body as stream
-        const originalBody = originalResponse.body;
-        // get response headers
-        const head = new Headers(originalResponse.headers);
-        head.delete('content-security-policy');
-        head.delete('content-security-policy-report-only');
-        head.delete('clear-site-data');
-        head.delete('Cache-Control');
-        head.delete('Content-Encoding');
-        const headObj = {};
-        for (let [k, v] of head) {
-            headObj[k] = v;
-        }
-        // send response headers to client
-        res.header(headObj);
-        // send response status to client
-        res.status(originalResponse.status);
-        // send response body to client
-        await pipeline(originalResponse.body, res);
-    }
-    catch (err) {
-        if (err?.name === 'AbortError') {
-            if (!res.headersSent) {
-                res.status(504).send({
-                    error: timeoutMs
-                        ? `Proxy request timed out after ${timeoutMs}ms`
-                        : 'Proxy request aborted'
-                });
-            } else {
-                res.end();
-            }
-            return;
-        }
-        next(err);
-        return;
-    } finally {
-        timeout.cleanup();
-    }
-}
-
-async function getSionywAccessToken() {
-    if(state.accessTokenCache.token && Date.now() < state.accessTokenCache.expiry){
-        return state.accessTokenCache.token;
-    }
-    //Schema of the client data file
-    // {
-    //     refresh_token: string;
-    //     client_id: string;
-    //     client_secret: string;
-    // }
-    
-    const clientDataPath = path.join(process.cwd(), 'save', '__sionyw_client_data.json');
-    let refreshToken = ''
-    let clientId = ''
-    let clientSecret = ''
-    if(!existsSync(clientDataPath)){
-        throw new Error('No Sionyw client data found');
-    }
-    const clientDataRaw = readFileSync(clientDataPath, 'utf-8');
-    const clientData = JSON.parse(clientDataRaw);
-    refreshToken = clientData.refresh_token;
-    clientId = clientData.client_id;
-    clientSecret = clientData.client_secret;
-
-    //Oauth Refresh Token Flow
-    
-    const tokenResponse = await fetch('account.sionyw.com/account/api/oauth/token', {
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/x-www-form-urlencoded'
-        },
-        body: new URLSearchParams({
-            grant_type: 'refresh_token',
-            refresh_token: refreshToken,
-            client_id: clientId,
-            client_secret: clientSecret
-        })
-    })
-
-    if(!tokenResponse.ok){
-        throw new Error('Failed to refresh Sionyw access token');
-    }
-
-    const tokenData = await tokenResponse.json();
-
-    //Update the refresh token in the client data file
-    if(tokenData.refresh_token && tokenData.refresh_token !== refreshToken){
-        clientData.refresh_token = tokenData.refresh_token;
-        writeFileSync(clientDataPath, JSON.stringify(clientData), 'utf-8');
-    }
-
-    state.accessTokenCache.token = tokenData.access_token;
-    state.accessTokenCache.expiry = Date.now() + (tokenData.expires_in * 1000) - (5 * 60 * 1000); //5 minutes early
-
-    return tokenData.access_token;
-}
-
-
-async function hubProxyFunc(req, res) {
-    const excludedHeaders = [
-        'content-encoding',
-        'content-length',
-        'transfer-encoding'
-    ];
-
-    try {
-        let externalURL = '';
-
-        const pathHeader = req.headers['x-risu-node-path'];
-        if (pathHeader) {
-            const decodedPath = decodeURIComponent(pathHeader);
-            externalURL = decodedPath;
-        } else {
-            const pathAndQuery = req.originalUrl.replace(/^\/hub-proxy/, '');
-            externalURL = hubURL + pathAndQuery;
-        }
-        
-        const headersToSend = { ...req.headers };
-        delete headersToSend.host;
-        delete headersToSend.connection;
-        delete headersToSend['content-length'];
-        delete headersToSend['x-risu-node-path'];
-
-        const hubOrigin = new URL(hubURL).origin;
-        headersToSend.origin = hubOrigin;
-
-        //if Authorization header is "Server-Auth, set the token to be Server-Auth
-        if(headersToSend['Authorization'] === 'X-Node-Server-Auth'){
-            //this requires password auth
-            if(!await checkAuth(req, res)){
-                return;
-            }
-
-            headersToSend['Authorization'] = "Bearer " + await getSionywAccessToken();
-            delete headersToSend['risu-auth'];
-        }
-        
-        
-        const response = await fetch(externalURL, {
-            method: req.method,
-            headers: headersToSend,
-            body: req.method !== 'GET' && req.method !== 'HEAD' ? req.body : undefined,
-            redirect: 'manual',
-            duplex: 'half'
-        });
-        
-        for (const [key, value] of response.headers.entries()) {
-            // Skip encoding-related headers to prevent double decoding
-            if (excludedHeaders.includes(key.toLowerCase())) {
-                continue;
-            }
-            res.setHeader(key, value);
-        }
-        res.status(response.status);
-
-        if (response.status >= 300 && response.status < 400 && response.headers.get('location')) {
-            const redirectUrl = response.headers.get('location');
-            const newHeaders = { ...headersToSend };
-            const redirectResponse = await fetch(redirectUrl, {
-                method: req.method,
-                headers: newHeaders,
-                body: req.method !== 'GET' && req.method !== 'HEAD' ? req.body : undefined,
-                redirect: 'manual',
-                duplex: 'half'
-            });
-            for (const [key, value] of redirectResponse.headers.entries()) {
-                if (excludedHeaders.includes(key.toLowerCase())) {
-                    continue;
-                }
-                res.setHeader(key, value);
-            }
-            res.status(redirectResponse.status);
-            if (redirectResponse.body) {
-                await pipeline(redirectResponse.body, res);
-            } else {
-                res.end();
-            }
-            return;
-        }
-        
-        if (response.body) {
-            await pipeline(response.body, res);
-        } else {
-            res.end();
-        }
-        
-    } catch (error) {
-        console.error("[Hub Proxy] Error:", error);
-        if (!res.headersSent) {
-            res.status(502).send({ error: 'Proxy request failed: ' + error.message });
-        } else {
-            res.end();
-        }
-    }
-}
-
-app.get('/proxy', authenticatedRouteLimiter, reverseProxyFunc_get);
-app.get('/proxy2', authenticatedRouteLimiter, reverseProxyFunc_get);
-app.get('/hub-proxy/*', authenticatedRouteLimiter, hubProxyFunc);
-
-app.post('/proxy', authenticatedRouteLimiter, reverseProxyFunc);
-app.post('/proxy2', authenticatedRouteLimiter, reverseProxyFunc);
-app.post('/hub-proxy/*', authenticatedRouteLimiter, hubProxyFunc);
+registerProxyRoutes(app, {
+    state,
+    authenticatedRouteLimiter,
+    authHelpers,
+});
 app.post('/proxy-stream-jobs', authenticatedRouteLimiter, async (req, res) => {
     if (!await checkProxyAuth(req, res)) {
         return;
